LayoutTests/imported/w3c/web-platform-tests/css/css-ui/animation/caret-color-interpolation-expected.txt


PASS CSS Transitions: property <caret-color> from neutral to [green] at (-0.3) should be [rgb(255, 255, 0)] 
PASS CSS Transitions: property <caret-color> from neutral to [green] at (0) should be [rgb(255, 255, 0)] 
PASS CSS Transitions: property <caret-color> from neutral to [green] at (0.3) should be [rgb(179, 217, 0)] 
PASS CSS Transitions: property <caret-color> from neutral to [green] at (0.6) should be [rgb(102, 179, 0)] 
PASS CSS Transitions: property <caret-color> from neutral to [green] at (1) should be [rgb(0, 128, 0)] 
PASS CSS Transitions: property <caret-color> from neutral to [green] at (1.5) should be [rgb(0, 65, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from neutral to [green] at (-0.3) should be [rgb(255, 255, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from neutral to [green] at (0) should be [rgb(255, 255, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from neutral to [green] at (0.3) should be [rgb(179, 217, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from neutral to [green] at (0.6) should be [rgb(102, 179, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from neutral to [green] at (1) should be [rgb(0, 128, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from neutral to [green] at (1.5) should be [rgb(0, 65, 0)] 
PASS CSS Animations: property <caret-color> from neutral to [green] at (-0.3) should be [rgb(255, 255, 0)] 
PASS CSS Animations: property <caret-color> from neutral to [green] at (0) should be [rgb(255, 255, 0)] 
PASS CSS Animations: property <caret-color> from neutral to [green] at (0.3) should be [rgb(179, 217, 0)] 
PASS CSS Animations: property <caret-color> from neutral to [green] at (0.6) should be [rgb(102, 179, 0)] 
PASS CSS Animations: property <caret-color> from neutral to [green] at (1) should be [rgb(0, 128, 0)] 
PASS CSS Animations: property <caret-color> from neutral to [green] at (1.5) should be [rgb(0, 65, 0)] 
PASS Web Animations: property <caret-color> from neutral to [green] at (-0.3) should be [rgb(255, 255, 0)] 
PASS Web Animations: property <caret-color> from neutral to [green] at (0) should be [rgb(255, 255, 0)] 
PASS Web Animations: property <caret-color> from neutral to [green] at (0.3) should be [rgb(179, 217, 0)] 
FAIL Web Animations: property <caret-color> from neutral to [green] at (0.6) should be [rgb(102, 179, 0)] assert_equals: expected "rgb ( 102 , 179 , 0 ) " but got "rgb ( 72 , 164 , 0 ) "
PASS Web Animations: property <caret-color> from neutral to [green] at (1) should be [rgb(0, 128, 0)] 
FAIL Web Animations: property <caret-color> from neutral to [green] at (1.5) should be [rgb(0, 65, 0)] assert_equals: expected "rgb ( 0 , 65 , 0 ) " but got "rgb ( 0 , 128 , 0 ) "
FAIL CSS Transitions: property <caret-color> from [initial] to [green] at (-0.3) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Transitions: property <caret-color> from [initial] to [green] at (0) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Transitions: property <caret-color> from [initial] to [green] at (0.3) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 126 , 0 , 0.3 ) "
FAIL CSS Transitions: property <caret-color> from [initial] to [green] at (0.5) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 128 , 0 , 0.5 ) "
FAIL CSS Transitions: property <caret-color> from [initial] to [green] at (0.6) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 129 , 0 , 0.6 ) "
PASS CSS Transitions: property <caret-color> from [initial] to [green] at (1) should be [green] 
FAIL CSS Transitions: property <caret-color> from [initial] to [green] at (1.5) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgb ( 0 , 192 , 0 ) "
FAIL CSS Transitions with transition: all: property <caret-color> from [initial] to [green] at (-0.3) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Transitions with transition: all: property <caret-color> from [initial] to [green] at (0) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Transitions with transition: all: property <caret-color> from [initial] to [green] at (0.3) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 126 , 0 , 0.3 ) "
FAIL CSS Transitions with transition: all: property <caret-color> from [initial] to [green] at (0.5) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 128 , 0 , 0.5 ) "
FAIL CSS Transitions with transition: all: property <caret-color> from [initial] to [green] at (0.6) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 129 , 0 , 0.6 ) "
PASS CSS Transitions with transition: all: property <caret-color> from [initial] to [green] at (1) should be [green] 
FAIL CSS Transitions with transition: all: property <caret-color> from [initial] to [green] at (1.5) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgb ( 0 , 192 , 0 ) "
FAIL CSS Animations: property <caret-color> from [initial] to [green] at (-0.3) should be [initial] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Animations: property <caret-color> from [initial] to [green] at (0) should be [initial] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Animations: property <caret-color> from [initial] to [green] at (0.3) should be [initial] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 126 , 0 , 0.3 ) "
FAIL CSS Animations: property <caret-color> from [initial] to [green] at (0.5) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 128 , 0 , 0.5 ) "
FAIL CSS Animations: property <caret-color> from [initial] to [green] at (0.6) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 129 , 0 , 0.6 ) "
PASS CSS Animations: property <caret-color> from [initial] to [green] at (1) should be [green] 
FAIL CSS Animations: property <caret-color> from [initial] to [green] at (1.5) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgb ( 0 , 192 , 0 ) "
FAIL Web Animations: property <caret-color> from [initial] to [green] at (-0.3) should be [initial] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL Web Animations: property <caret-color> from [initial] to [green] at (0) should be [initial] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL Web Animations: property <caret-color> from [initial] to [green] at (0.3) should be [initial] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 126 , 0 , 0.3 ) "
FAIL Web Animations: property <caret-color> from [initial] to [green] at (0.5) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 128 , 0 , 0.5 ) "
FAIL Web Animations: property <caret-color> from [initial] to [green] at (0.6) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 129 , 0 , 0.6 ) "
PASS Web Animations: property <caret-color> from [initial] to [green] at (1) should be [green] 
FAIL Web Animations: property <caret-color> from [initial] to [green] at (1.5) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgb ( 0 , 192 , 0 ) "
FAIL CSS Transitions: property <caret-color> from [auto] to [green] at (-0.3) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Transitions: property <caret-color> from [auto] to [green] at (0) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Transitions: property <caret-color> from [auto] to [green] at (0.3) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 126 , 0 , 0.3 ) "
FAIL CSS Transitions: property <caret-color> from [auto] to [green] at (0.5) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 128 , 0 , 0.5 ) "
FAIL CSS Transitions: property <caret-color> from [auto] to [green] at (0.6) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 129 , 0 , 0.6 ) "
PASS CSS Transitions: property <caret-color> from [auto] to [green] at (1) should be [green] 
FAIL CSS Transitions: property <caret-color> from [auto] to [green] at (1.5) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgb ( 0 , 192 , 0 ) "
FAIL CSS Transitions with transition: all: property <caret-color> from [auto] to [green] at (-0.3) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Transitions with transition: all: property <caret-color> from [auto] to [green] at (0) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Transitions with transition: all: property <caret-color> from [auto] to [green] at (0.3) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 126 , 0 , 0.3 ) "
FAIL CSS Transitions with transition: all: property <caret-color> from [auto] to [green] at (0.5) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 128 , 0 , 0.5 ) "
FAIL CSS Transitions with transition: all: property <caret-color> from [auto] to [green] at (0.6) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 129 , 0 , 0.6 ) "
PASS CSS Transitions with transition: all: property <caret-color> from [auto] to [green] at (1) should be [green] 
FAIL CSS Transitions with transition: all: property <caret-color> from [auto] to [green] at (1.5) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgb ( 0 , 192 , 0 ) "
FAIL CSS Animations: property <caret-color> from [auto] to [green] at (-0.3) should be [auto] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Animations: property <caret-color> from [auto] to [green] at (0) should be [auto] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Animations: property <caret-color> from [auto] to [green] at (0.3) should be [auto] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 126 , 0 , 0.3 ) "
FAIL CSS Animations: property <caret-color> from [auto] to [green] at (0.5) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 128 , 0 , 0.5 ) "
FAIL CSS Animations: property <caret-color> from [auto] to [green] at (0.6) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 129 , 0 , 0.6 ) "
PASS CSS Animations: property <caret-color> from [auto] to [green] at (1) should be [green] 
FAIL CSS Animations: property <caret-color> from [auto] to [green] at (1.5) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgb ( 0 , 192 , 0 ) "
FAIL Web Animations: property <caret-color> from [auto] to [green] at (-0.3) should be [auto] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL Web Animations: property <caret-color> from [auto] to [green] at (0) should be [auto] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL Web Animations: property <caret-color> from [auto] to [green] at (0.3) should be [auto] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 126 , 0 , 0.3 ) "
FAIL Web Animations: property <caret-color> from [auto] to [green] at (0.5) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 128 , 0 , 0.5 ) "
FAIL Web Animations: property <caret-color> from [auto] to [green] at (0.6) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgba ( 0 , 129 , 0 , 0.6 ) "
PASS Web Animations: property <caret-color> from [auto] to [green] at (1) should be [green] 
FAIL Web Animations: property <caret-color> from [auto] to [green] at (1.5) should be [green] assert_equals: expected "rgb ( 0 , 128 , 0 ) " but got "rgb ( 0 , 192 , 0 ) "
FAIL CSS Transitions: property <caret-color> from [currentColor] to [green] at (-0.3) should be [rgb(0, 0, 0)] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Transitions: property <caret-color> from [currentColor] to [green] at (0) should be [rgb(0, 0, 0)] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Transitions: property <caret-color> from [currentColor] to [green] at (0.3) should be [rgb(0, 38, 0)] assert_equals: expected "rgb ( 0 , 38 , 0 ) " but got "rgba ( 0 , 126 , 0 , 0.3 ) "
FAIL CSS Transitions: property <caret-color> from [currentColor] to [green] at (0.6) should be [rgb(0, 77, 0)] assert_equals: expected "rgb ( 0 , 77 , 0 ) " but got "rgba ( 0 , 129 , 0 , 0.6 ) "
PASS CSS Transitions: property <caret-color> from [currentColor] to [green] at (1) should be [rgb(0, 128, 0)] 
PASS CSS Transitions: property <caret-color> from [currentColor] to [green] at (1.5) should be [rgb(0, 192, 0)] 
FAIL CSS Transitions with transition: all: property <caret-color> from [currentColor] to [green] at (-0.3) should be [rgb(0, 0, 0)] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Transitions with transition: all: property <caret-color> from [currentColor] to [green] at (0) should be [rgb(0, 0, 0)] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Transitions with transition: all: property <caret-color> from [currentColor] to [green] at (0.3) should be [rgb(0, 38, 0)] assert_equals: expected "rgb ( 0 , 38 , 0 ) " but got "rgba ( 0 , 126 , 0 , 0.3 ) "
FAIL CSS Transitions with transition: all: property <caret-color> from [currentColor] to [green] at (0.6) should be [rgb(0, 77, 0)] assert_equals: expected "rgb ( 0 , 77 , 0 ) " but got "rgba ( 0 , 129 , 0 , 0.6 ) "
PASS CSS Transitions with transition: all: property <caret-color> from [currentColor] to [green] at (1) should be [rgb(0, 128, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from [currentColor] to [green] at (1.5) should be [rgb(0, 192, 0)] 
FAIL CSS Animations: property <caret-color> from [currentColor] to [green] at (-0.3) should be [rgb(0, 0, 0)] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Animations: property <caret-color> from [currentColor] to [green] at (0) should be [rgb(0, 0, 0)] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL CSS Animations: property <caret-color> from [currentColor] to [green] at (0.3) should be [rgb(0, 38, 0)] assert_equals: expected "rgb ( 0 , 38 , 0 ) " but got "rgba ( 0 , 126 , 0 , 0.3 ) "
FAIL CSS Animations: property <caret-color> from [currentColor] to [green] at (0.6) should be [rgb(0, 77, 0)] assert_equals: expected "rgb ( 0 , 77 , 0 ) " but got "rgba ( 0 , 129 , 0 , 0.6 ) "
PASS CSS Animations: property <caret-color> from [currentColor] to [green] at (1) should be [rgb(0, 128, 0)] 
PASS CSS Animations: property <caret-color> from [currentColor] to [green] at (1.5) should be [rgb(0, 192, 0)] 
FAIL Web Animations: property <caret-color> from [currentColor] to [green] at (-0.3) should be [rgb(0, 0, 0)] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL Web Animations: property <caret-color> from [currentColor] to [green] at (0) should be [rgb(0, 0, 0)] assert_equals: expected "rgb ( 0 , 0 , 0 ) " but got "rgba ( 0 , 0 , 0 , 0 ) "
FAIL Web Animations: property <caret-color> from [currentColor] to [green] at (0.3) should be [rgb(0, 38, 0)] assert_equals: expected "rgb ( 0 , 38 , 0 ) " but got "rgba ( 0 , 126 , 0 , 0.3 ) "
FAIL Web Animations: property <caret-color> from [currentColor] to [green] at (0.6) should be [rgb(0, 77, 0)] assert_equals: expected "rgb ( 0 , 77 , 0 ) " but got "rgba ( 0 , 129 , 0 , 0.6 ) "
PASS Web Animations: property <caret-color> from [currentColor] to [green] at (1) should be [rgb(0, 128, 0)] 
PASS Web Animations: property <caret-color> from [currentColor] to [green] at (1.5) should be [rgb(0, 192, 0)] 
PASS CSS Transitions: property <caret-color> from [inherit] to [green] at (-0.3) should be [rgb(0, 0, 255)] 
PASS CSS Transitions: property <caret-color> from [inherit] to [green] at (0) should be [rgb(0, 0, 255)] 
PASS CSS Transitions: property <caret-color> from [inherit] to [green] at (0.3) should be [rgb(0, 38, 179)] 
PASS CSS Transitions: property <caret-color> from [inherit] to [green] at (0.6) should be [rgb(0, 77, 102)] 
PASS CSS Transitions: property <caret-color> from [inherit] to [green] at (1) should be [rgb(0, 128, 0)] 
PASS CSS Transitions: property <caret-color> from [inherit] to [green] at (1.5) should be [rgb(0, 192, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from [inherit] to [green] at (-0.3) should be [rgb(0, 0, 255)] 
PASS CSS Transitions with transition: all: property <caret-color> from [inherit] to [green] at (0) should be [rgb(0, 0, 255)] 
PASS CSS Transitions with transition: all: property <caret-color> from [inherit] to [green] at (0.3) should be [rgb(0, 38, 179)] 
PASS CSS Transitions with transition: all: property <caret-color> from [inherit] to [green] at (0.6) should be [rgb(0, 77, 102)] 
PASS CSS Transitions with transition: all: property <caret-color> from [inherit] to [green] at (1) should be [rgb(0, 128, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from [inherit] to [green] at (1.5) should be [rgb(0, 192, 0)] 
PASS CSS Animations: property <caret-color> from [inherit] to [green] at (-0.3) should be [rgb(0, 0, 255)] 
PASS CSS Animations: property <caret-color> from [inherit] to [green] at (0) should be [rgb(0, 0, 255)] 
PASS CSS Animations: property <caret-color> from [inherit] to [green] at (0.3) should be [rgb(0, 38, 179)] 
PASS CSS Animations: property <caret-color> from [inherit] to [green] at (0.6) should be [rgb(0, 77, 102)] 
PASS CSS Animations: property <caret-color> from [inherit] to [green] at (1) should be [rgb(0, 128, 0)] 
PASS CSS Animations: property <caret-color> from [inherit] to [green] at (1.5) should be [rgb(0, 192, 0)] 
FAIL Web Animations: property <caret-color> from [inherit] to [green] at (-0.3) should be [rgb(0, 0, 255)] assert_equals: expected "rgb ( 0 , 0 , 255 ) " but got "rgb ( 255 , 255 , 0 ) "
FAIL Web Animations: property <caret-color> from [inherit] to [green] at (0) should be [rgb(0, 0, 255)] assert_equals: expected "rgb ( 0 , 0 , 255 ) " but got "rgb ( 255 , 255 , 0 ) "
FAIL Web Animations: property <caret-color> from [inherit] to [green] at (0.3) should be [rgb(0, 38, 179)] assert_equals: expected "rgb ( 0 , 38 , 179 ) " but got "rgb ( 179 , 217 , 0 ) "
FAIL Web Animations: property <caret-color> from [inherit] to [green] at (0.6) should be [rgb(0, 77, 102)] assert_equals: expected "rgb ( 0 , 77 , 102 ) " but got "rgb ( 72 , 164 , 0 ) "
PASS Web Animations: property <caret-color> from [inherit] to [green] at (1) should be [rgb(0, 128, 0)] 
FAIL Web Animations: property <caret-color> from [inherit] to [green] at (1.5) should be [rgb(0, 192, 0)] assert_equals: expected "rgb ( 0 , 192 , 0 ) " but got "rgb ( 0 , 128 , 0 ) "
PASS CSS Transitions: property <caret-color> from [unset] to [green] at (-0.3) should be [rgb(0, 0, 255)] 
PASS CSS Transitions: property <caret-color> from [unset] to [green] at (0) should be [rgb(0, 0, 255)] 
PASS CSS Transitions: property <caret-color> from [unset] to [green] at (0.3) should be [rgb(0, 38, 179)] 
PASS CSS Transitions: property <caret-color> from [unset] to [green] at (0.6) should be [rgb(0, 77, 102)] 
PASS CSS Transitions: property <caret-color> from [unset] to [green] at (1) should be [rgb(0, 128, 0)] 
PASS CSS Transitions: property <caret-color> from [unset] to [green] at (1.5) should be [rgb(0, 192, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from [unset] to [green] at (-0.3) should be [rgb(0, 0, 255)] 
PASS CSS Transitions with transition: all: property <caret-color> from [unset] to [green] at (0) should be [rgb(0, 0, 255)] 
PASS CSS Transitions with transition: all: property <caret-color> from [unset] to [green] at (0.3) should be [rgb(0, 38, 179)] 
PASS CSS Transitions with transition: all: property <caret-color> from [unset] to [green] at (0.6) should be [rgb(0, 77, 102)] 
PASS CSS Transitions with transition: all: property <caret-color> from [unset] to [green] at (1) should be [rgb(0, 128, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from [unset] to [green] at (1.5) should be [rgb(0, 192, 0)] 
PASS CSS Animations: property <caret-color> from [unset] to [green] at (-0.3) should be [rgb(0, 0, 255)] 
PASS CSS Animations: property <caret-color> from [unset] to [green] at (0) should be [rgb(0, 0, 255)] 
PASS CSS Animations: property <caret-color> from [unset] to [green] at (0.3) should be [rgb(0, 38, 179)] 
PASS CSS Animations: property <caret-color> from [unset] to [green] at (0.6) should be [rgb(0, 77, 102)] 
PASS CSS Animations: property <caret-color> from [unset] to [green] at (1) should be [rgb(0, 128, 0)] 
PASS CSS Animations: property <caret-color> from [unset] to [green] at (1.5) should be [rgb(0, 192, 0)] 
FAIL Web Animations: property <caret-color> from [unset] to [green] at (-0.3) should be [rgb(0, 0, 255)] assert_equals: expected "rgb ( 0 , 0 , 255 ) " but got "rgb ( 255 , 255 , 0 ) "
FAIL Web Animations: property <caret-color> from [unset] to [green] at (0) should be [rgb(0, 0, 255)] assert_equals: expected "rgb ( 0 , 0 , 255 ) " but got "rgb ( 255 , 255 , 0 ) "
FAIL Web Animations: property <caret-color> from [unset] to [green] at (0.3) should be [rgb(0, 38, 179)] assert_equals: expected "rgb ( 0 , 38 , 179 ) " but got "rgb ( 179 , 217 , 0 ) "
FAIL Web Animations: property <caret-color> from [unset] to [green] at (0.6) should be [rgb(0, 77, 102)] assert_equals: expected "rgb ( 0 , 77 , 102 ) " but got "rgb ( 72 , 164 , 0 ) "
PASS Web Animations: property <caret-color> from [unset] to [green] at (1) should be [rgb(0, 128, 0)] 
FAIL Web Animations: property <caret-color> from [unset] to [green] at (1.5) should be [rgb(0, 192, 0)] assert_equals: expected "rgb ( 0 , 192 , 0 ) " but got "rgb ( 0 , 128 , 0 ) "
PASS CSS Transitions: property <caret-color> from [black] to [orange] at (-0.3) should be [rgb(0, 0, 0)] 
PASS CSS Transitions: property <caret-color> from [black] to [orange] at (0) should be [rgb(0, 0, 0)] 
PASS CSS Transitions: property <caret-color> from [black] to [orange] at (0.3) should be [rgb(77, 50, 0)] 
PASS CSS Transitions: property <caret-color> from [black] to [orange] at (0.6) should be [rgb(153, 99, 0)] 
PASS CSS Transitions: property <caret-color> from [black] to [orange] at (1) should be [rgb(255, 165, 0)] 
PASS CSS Transitions: property <caret-color> from [black] to [orange] at (1.5) should be [rgb(255, 248, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from [black] to [orange] at (-0.3) should be [rgb(0, 0, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from [black] to [orange] at (0) should be [rgb(0, 0, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from [black] to [orange] at (0.3) should be [rgb(77, 50, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from [black] to [orange] at (0.6) should be [rgb(153, 99, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from [black] to [orange] at (1) should be [rgb(255, 165, 0)] 
PASS CSS Transitions with transition: all: property <caret-color> from [black] to [orange] at (1.5) should be [rgb(255, 248, 0)] 
PASS CSS Animations: property <caret-color> from [black] to [orange] at (-0.3) should be [rgb(0, 0, 0)] 
PASS CSS Animations: property <caret-color> from [black] to [orange] at (0) should be [rgb(0, 0, 0)] 
PASS CSS Animations: property <caret-color> from [black] to [orange] at (0.3) should be [rgb(77, 50, 0)] 
PASS CSS Animations: property <caret-color> from [black] to [orange] at (0.6) should be [rgb(153, 99, 0)] 
PASS CSS Animations: property <caret-color> from [black] to [orange] at (1) should be [rgb(255, 165, 0)] 
PASS CSS Animations: property <caret-color> from [black] to [orange] at (1.5) should be [rgb(255, 248, 0)] 
PASS Web Animations: property <caret-color> from [black] to [orange] at (-0.3) should be [rgb(0, 0, 0)] 
PASS Web Animations: property <caret-color> from [black] to [orange] at (0) should be [rgb(0, 0, 0)] 
PASS Web Animations: property <caret-color> from [black] to [orange] at (0.3) should be [rgb(77, 50, 0)] 
PASS Web Animations: property <caret-color> from [black] to [orange] at (0.6) should be [rgb(153, 99, 0)] 
PASS Web Animations: property <caret-color> from [black] to [orange] at (1) should be [rgb(255, 165, 0)] 
PASS Web Animations: property <caret-color> from [black] to [orange] at (1.5) should be [rgb(255, 248, 0)] 
